requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic[email] 
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
alembic==1.13.2
psycopg2-binary==2.9.10
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose==3.3.0
aio-pika==9.4.3